Anime API

Contents

Anime API

Contents

Index

Index

Anime.Anime

Anime.addweightcols

Anime.attenuate

Anime.bandpass

Anime.compute_meandelays

Anime.compute_skynoise

Anime.compute_turbulence

Anime.computeweights!

Anime.elevationangle

Anime.gentimeseries!

Anime.gentimeseries!

Anime.gentimeseries!

Anime.instrumentalpol

Anime.loadms

Anime.makecasaanttable

Anime.msfromconfig

Anime.msfromuvfits

Anime.msfromvex

Anime.parallacticangle

Anime.plotbandpass

Anime.plotpointingerrors

Anime.plotstationgains

Anime.plotvis

Anime.pointing

Anime.postprocessms

Anime.run_wsclean

Anime.setup_polarization

Anime.stationgains

Anime.thermalnoise

Anime.troposphere

Anime.Anime — Module

Anime: Atmospheric aNd Instrumental Models in the Measurement Equation

Anime.addweightcols — Method

addweightcols(msname::String, mode::String, sigmaspec::Bool, weightspec::Bool)

Add WEIGHTSPECTRUM and SIGMASPECTRUM columns to the MS

Anime.attenuate — Method

attenuate(obs::CjlObservation, transmission::Array{Float64, 3})

Compute attenuation due to troposphere

Anime.bandpass — Method

bandpass(bandpassfile::String, data::Array{Complex{Float32},4}, stationinfo::DataFrame, rngcorrupt::AbstractRNG,
antenna1::Vector{Int32}, antenna2::Vector{Int32}, numchan::Int64, chanfreqvec::Vector{Float64}; h5file::String="")

Compute the bandpass model and apply to data. The actual numerical values are serialized as HDF5.

Anime.compute_meandelays — Method

compute_meandelays(obs::CjlObservation, atmdf::DataFrame, elevationmatrix::Array{Float64, 2}, g::HDF5.Group)

Compute mean delays

Anime.compute_skynoise — Method

compute_skynoise(obs::CjlObservation, atmdf::DataFrame, transmission::Array{Float64, 3}, g::HDF5.Group)

Compute sky noise contribution

Anime.compute_turbulence — Method

compute_turbulence(obs::CjlObservation, atmdf::DataFrame, elevationmatrix::Array{Float64, 2}, g::HDF5.Group)

Compute turbulent phases

Anime.computeweights! — Method

computeweights!(totalrmsspec::Array{Float32, 4}, totalwtspec::Array{Float32, 4}, obs::CjlObservation; h5file::String="")

Compute total rms (sigma) values from independent noise sources

Anime.elevationangle — Method

elevationangle(times::Vector{Float64}, phasedir::Array{Float64,2}, stationinfo::DataFrame, pos::Array{Float64, 2})

Compute elevation angle for all stations for all times.

Anime.gentimeseries! — Method

gentimeseries!(series::Vector{ComplexF32}, mode::String, location::ComplexF32, scale::Float32, driftrate::Float32, nsamples::Int64, rng::AbstractRNG)

Generate a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.

Anime.gentimeseries! — Method

gentimeseries!(series::Vector{Float32}, mode::String, location::Float32, scale::Float32, driftrate::Float32, nsamples::Int64, rng::AbstractRNG)

Generate a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.

Anime.gentimeseries! — Method

gentimeseries!(series::Vector{Float64}, mode::String, location::Float64, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)

Generate a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.

Anime.instrumentalpol — Method

instrumentalpol(scanno::Vector{Int32}, times::Vector{Float64}, stationinfo::DataFrame, phasedir::Array{Float64,2},
pos::Array{Float64, 2}, data::Array{Complex{Float32},4}, numchan::Int64, polframe::String, polmode::String,
antenna1::Vector{Int32}, antenna2::Vector{Int32}, exposure::Float64, rngcorrupt::AbstractRNG; h5file::String="")

Compute instrumental polarization (leakage, or "D-Jones" terms) and apply to data. The actual numerical values are serialized as HDF5.

Anime.loadms — Method

loadms(msname::String, stations::String, corruptseed::Int64, tropseed::Int64, tropwetonly::Bool, correff::Float64, tropattenuate::Bool,
tropskynoise::Bool, tropmeandelays::Bool, tropturbulence::Bool, polframe::String, polmode::String, ptginterval::Float64, ptgmode::String,
stationgainsmode::String, bandpassfile::String; delim::String=",", ignorerepeated::Bool=false)

Load data and metadata from MS and station table and return a CjlObservation object.

Anime.makecasaanttable — Method

makecasaanttable(stations::String, casaanttemplate::String; delim::String=",", ignorerepeated::Bool=false)

Generate a CASA antenna table from CSV station info file in the current working directory.

Anime.msfromconfig — Method

msfromconfig(msname::String, mscreationmode::String, stations::String, casaanttemplate::String, spw_centrefreq::Array{Float64, 1}, 
spw_bw::Array{Float64, 1}, spw_channels::Array{Int64, 1}, sourcedict::Dict{String, Any}, starttime::String, exposure::Float64, scans::Int64,
scanlengths::Array{Float64, 1}, scanlag::Float64; autocorr::Bool=false, telescopename::String="VLBA", feed::String="perfect R L",
shadowlimit::Float64=1e-6, elevationlimit::String="10deg", stokes::String="RR RL LR LL", delim::String=",", ignorerepeated::Bool=false)

Generate measurement set from scratch from input observation parameters

Anime.msfromuvfits — Method

msfromuvfits(uvfits::String, msname::String, mscreationmode::String, stations::String; delim::String=",", ignorerepeated::Bool=false)

Generate MS from existing UVFITS file

Anime.msfromvex — Method

msfromvex()

Extract config parameters from VEX schedule and call msfromconfig() to generate MS

Anime.parallacticangle — Method

parallacticangle(times::Vector{Float64}, phasedir::Array{Float64,2}, stationinfo::DataFrame, pos::Array{Float64,2})

Compute parallactic angle for all stations for all times.

Anime.plotbandpass — Method

plotbandpass(h5file::String, stationnames::Vector{String3}, chanfreqvec::Vector{Float64}; saveas="bandpassgains.png")

Plot bandpass gains against time

Anime.plotpointingerrors — Method

plotpointingerrors(h5file::String, scanno::Vector{Int32}, stationnames::Vector{String3})

Plot pointing errors

Anime.plotstationgains — Method

plotstationgains(h5file::String, scanno::Vector{Int32}, times::Vector{Float64}, stationnames::Vector{String3}; saveas="stationgainsvstime.png")

Plot station gains against time

Anime.plotvis — Method

plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Array{Bool,4}, data::Array{Complex{Float32},4},
numchan::Int64, times::Vector{Float64}; saveprefix="data_")

Generate various complex visibility plots

Anime.pointing — Method

pointing(stationinfo::DataFrame, scanno::Vector{Int32}, chanfreqvec::Vector{Float64}, ptgint::Float64, ptgmode::String,
exposure::Float64, times::Vector{Float64}, rngcorrupt::AbstractRNG, antenna1::Vector{Int32}, antenna2::Vector{Int32},
data::Array{Complex{Float32},4}, numchan::Int64; h5file::String="")

Compute the pointing model and apply to data. The actual numerical values are serialized as HDF5.

Anime.postprocessms — Method

postprocessms(obs::CjlObservation; h5file::String="")

Add weights and sigma matrices, update data matrix, reshape them all and write to MS

Anime.run_wsclean — Method

run_wsclean(msname::String, fitsdir::String, polarized::Bool, channelgroups::Int64, osfactor::Int64)

Compute source coherency using WSClean

Anime.setup_polarization — Method

setup_polarization(msname::String, stationtable::String)

Set up all polarization subtables in the MS based on the station feed types

Anime.stationgains — Method

stationgains(scanno::Vector{Int32}, times::Vector{Float64}, exposure::Float64, data::Array{Complex{Float32},4},
stationinfo::DataFrame, mode::String, rngcorrupt::AbstractRNG, antenna1::Vector{Int32}, antenna2::Vector{Int32}, numchan::Int64; h5file::String="")

Compute time-variable station gains and apply to data. The actual numerical values are serialized as HDF5.

Anime.thermalnoise — Method

thermalnoise(times::Vector{Float64}, antenna1::Vector{Int32}, antenna2::Vector{Int32}, data::Array{Complex{Float32},4}, correff::Float64,
exposure::Float64, chanwidth::Float64, rngcorrupt::AbstractRNG, sefd::Vector{Float64}; h5file::String="")

Compute per-baseline thermal noise in visibility domain and apply to data. The actual numerical values are serialized as HDF5.

Anime.troposphere — Method

troposphere(obs::CjlObservation; h5file::String="")

Compute various tropospheric effects and apply to data. The actual numerical values are serialized as HDF5.