Anime API

Contents

Anime API

Contents

Index

Index

Anime.addcorruptions

Anime.bandpass

Anime.elevationangle

Anime.generatems

Anime.gentimeseries!

Anime.gentimeseries!

Anime.gentimeseries!

Anime.instrumentalpol

Anime.loadobs

Anime.parallacticangle

Anime.pointing

Anime.predict_visibilities

Anime.stationgains

Anime.thermalnoise

Anime.troposphere

Anime.addcorruptions — Method

addcorruptions(obs::CjlObservation)

This function is a wrapper that applies the instrument model to the source coherency.

Anime.bandpass — Method

bandpass(obs::CjlObservation)

This function computes the bandpass model and applies it to the data. The actual numerical values are serialized.

Anime.elevationangle — Method

elevationangle(obs::CjlObservation)

This function computes the elevation angle for all stations for all times.

Anime.generatems — Method

generatems(config::String; delim::String=",", ignorerepeated::Bool=false)

This calls the appropriate MS creation function based on the input parameters in the config file.

Anime.gentimeseries! — Method

gentimeseries!(series::Vector{ComplexF32}, mode::String, location::ComplexF32, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)

This function generates a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.

Anime.gentimeseries! — Method

gentimeseries!(series::Vector{Float32}, mode::String, location::Float32, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)

This function generates a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.

Anime.gentimeseries! — Method

gentimeseries!(series::Vector{Float64}, mThis function generates a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.ode::String, location::Float64, scale::Float64, driftrate::Float64, nsamples::Int64, rng::AbstractRNG)

This function generates a complex-valued Gaussian process time-series of length nsamples with the given location, scale, and driftrate parameters.

Anime.instrumentalpol — Method

instrumentalpol(obs::CjlObservation)

This function computes the instrumental polarization (leakage, or "D-Jones" terms) and applies it to the data. The actual numerical values are serialized.

Anime.loadobs — Method

loadobs(config::String; delim::String=",", ignorerepeated::Bool=false)

This loads data and metadata from ms and station table and returns a CjlObservation object.

Anime.parallacticangle — Method

parallacticangle(obs::CjlObservation)

This function computes the parallactic angle for all stations for all times.

Anime.pointing — Method

pointing(obs::CjlObservation)

This function computes the pointing model and applies it to the data. The actual numerical values are serialized.

Anime.predict_visibilities — Method

predict_visibilities(config::String)

This predicts the uncorrupted visibilities with the parameters obtained from config file.

Anime.stationgains — Method

stationgains(obs::CjlObservation)

This function computes time-variable station gains and applies it to the data. The actual numerical values are serialized.

Anime.thermalnoise — Method

thermalnoise(obs::CjlObservation)

This function computes the per-baseline thermal noise in the visibility domain and applies it to the data. The actual numerical values are serialized.

Anime.troposphere — Method

troposphere(obs::CjlObservation)

This function computes various tropospheric effects and applies them to the data. The actual numerical values are serialized.